Separate registration from __init__
This reduces the dependencies of __init__ and solves a cycle which was
created during a previous refactoring.
