+++ cmk/gui/views/__init__.py
@@ -9,7 +9,6 @@ from typing import Any
 import cmk.gui.utils as utils
 import cmk.gui.visuals as visuals
 from cmk.gui.config import default_authorized_builtin_role_ids
-from cmk.gui.data_source import data_source_registry, register_data_sources
 from cmk.gui.i18n import _, _u
 from cmk.gui.pages import PageRegistry
 from cmk.gui.painter.v0 import painters
@@ -25,31 +24,14 @@ from cmk.gui.type_defs import Perfdata, PerfometerSpec, TranslatedMetrics, Visua
 from cmk.gui.view_utils import get_labels, render_labels, render_tag_groups
 from cmk.gui.visuals.type import VisualTypeRegistry
 
-from . import icon, inventory, perfometer
+from . import icon, inventory
 from .builtin_views import builtin_views
-from .command import (
-    command_group_registry,
-    command_registry,
-    register_command_groups,
-    register_commands,
-    register_legacy_command,
-)
-from .datasource_selection import page_select_datasource
-from .host_tag_plugins import register_tag_plugins
+from .command import register_legacy_command
 from .icon import Icon, icon_and_action_registry
-from .icon.page_ajax_popup_action_menu import ajax_popup_action_menu
 from .inventory import register_table_views_and_columns, update_paint_functions
-from .layout import layout_registry, register_layouts
-from .page_ajax_filters import AjaxInitialViewFilters
-from .page_ajax_reschedule import PageRescheduleCheck
-from .page_create_view import page_create_view
-from .page_edit_view import page_edit_view, PageAjaxCascadingRenderPainterParameters
-from .page_edit_views import page_edit_views
-from .page_show_view import page_show_view
 from .sorter import register_sorter, register_sorters, sorter_registry
 from .store import multisite_builtin_views
 from .view_choices import format_view_title
-from .visual_type import VisualTypeViews
 
 # TODO: Kept for compatibility with pre 1.6 plugins. Plugins will not be used anymore, but an error
 # will be displayed.
@@ -57,60 +39,6 @@ multisite_commands: list[dict[str, Any]] = []
 multisite_painters: dict[str, dict[str, Any]] = {}
 multisite_sorters: dict[str, Any] = {}
 multisite_icons_and_actions: dict[str, dict[str, Any]] = {}
-
-
-def register(
-    permission_section_registry: PermissionSectionRegistry,
-    page_registry: PageRegistry,
-    visual_type_registry: VisualTypeRegistry,
-    register_post_config_load_hook: Callable[[Callable[[], None]], None],
-) -> None:
-    register_post_config_load_hook(register_tag_plugins)
-
-    permission_section_registry.register(PermissionSectionViews)
-
-    page_registry.register_page("ajax_cascading_render_painer_parameters")(
-        PageAjaxCascadingRenderPainterParameters
-    )
-    page_registry.register_page("ajax_reschedule")(PageRescheduleCheck)
-    page_registry.register_page("ajax_initial_view_filters")(AjaxInitialViewFilters)
-    page_registry.register_page_handler("view", page_show_view)
-    page_registry.register_page_handler("create_view", page_select_datasource)
-    page_registry.register_page_handler("edit_view", page_edit_view)
-    page_registry.register_page_handler("edit_views", page_edit_views)
-    page_registry.register_page_handler("create_view_infos", page_create_view)
-    page_registry.register_page_handler("ajax_popup_action_menu", ajax_popup_action_menu)
-
-    visual_type_registry.register(VisualTypeViews)
-
-    register_layouts(layout_registry)
-    painters.register(painter_option_registry, painter_registry)
-    register_sorters(sorter_registry)
-    register_command_groups(command_group_registry)
-    register_commands(command_registry)
-    register_data_sources(data_source_registry)
-    perfometer.register(sorter_registry, painter_registry)
-    icon.register(
-        icon.icon_and_action_registry,
-        painter_registry,
-        permission_section_registry,
-        register_post_config_load_hook,
-    )
-    inventory.register(page_registry)
-
-
-class PermissionSectionViews(PermissionSection):
-    @property
-    def name(self) -> str:
-        return "view"
-
-    @property
-    def title(self) -> str:
-        return _("Views")
-
-    @property
-    def do_sort(self):
-        return True
 
 
 def load_plugins() -> None:
